comparison-with-state-of-the-art/eval/requirements.txt
altair==5.3.0
asttokens==2.4.1
attrs==23.2.0
bcrypt==4.1.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.2
cryptography==42.0.8
debugpy==1.8.2
decorator==5.1.1
docker==7.1.0
exceptiongroup==1.2.2
executing==2.0.1
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
paramiko==3.4.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pycparser==2.22
Pygments==2.18.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
six==1.16.0
stack-data==0.6.3
toolz==0.12.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vegafusion==1.6.9
vegafusion-python-embed==1.6.9
vl-convert-python==1.5.0
wcwidth==0.2.13
Werkzeug==3.0.3
